# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Multi-Auth Realtime API",
    "version": "1.0.0",
    "description": "Demonstrates Non-Auth, Basic, Digest, OAuth 1.0, OAuth 1.0a, OAuth 2.0 with realtime (WS, SSE, Socket.IO) - Single BaseUrl for all auth types"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Single BaseUrl for all auth types - Dedicated Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "digestAuth": {
        "type": "http",
        "scheme": "digest"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth1": {
        "type": "oauth2",
        "flows": {}
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:3000/api/oauth2/authorize",
            "tokenUrl": "http://localhost:3000/api/oauth2/token",
            "scopes": {
              "read": "read",
              "write": "write"
            }
          },
          "clientCredentials": {
            "tokenUrl": "http://localhost:3000/api/oauth2/token",
            "scopes": {
              "read": "read"
            }
          },
          "password": {
            "tokenUrl": "http://localhost:3000/api/oauth2/token",
            "scopes": {
              "read": "read",
              "write": "write"
            }
...
  },
  "paths": {
...
    "/api/basic/data": {
      "get": {
        "tags": [
          "Basic Auth"
        ],
        "summary": "Basic Auth data",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/basic/profile": {
      "get": {
        "tags": [
          "Basic Auth"
        ],
        "summary": "Basic profile",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/oauth1/data": {
      "get": {
        "tags": [
          "OAuth 1.0"
        ],
        "summary": "OAuth 1.0 data",
        "security": [
          {
            "oauth1": []
          }
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "OAuth header with HMAC-SHA1 signature",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/oauth1a/authorize": {
      "get": {
        "tags": [
          "OAuth 1.0a"
        ],
        "summary": "Authorize request token",
        "parameters": [
          {
            "name": "oauth_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "verifier"
          }
        }
      }
    },
    "/api/oauth1a/access_token": {
      "post": {
        "tags": [
          "OAuth 1.0a"
        ],
        "summary": "Exchange for access token",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "oauth_token": {
                    "type": "string"
                  },
                  "oauth_verifier": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "access token"
          }
        }
      }
    },
    "/api/oauth1a/data": {
      "get": {
        "tags": [
          "OAuth 1.0a"
        ],
        "summary": "OAuth1a data (needs access token)",
        "security": [
          {
            "oauth1": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/oauth2/authorize": {
      "get": {
        "tags": [
          "OAuth 2.0"
        ],
        "summary": "Authorization endpoint",
        "parameters": [
          {
            "name": "response_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "code"
              ]
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "code"
          }
        }
      }
    },
...
    "/api/oauth2/data": {
      "get": {
        "tags": [
          "OAuth 2.0"
        ],
        "summary": "OAuth2 protected data",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}